backend/requirements/test.txt
# Test dependencies go here.
-r base.txt


coverage==4.4.1
flake8==3.4.1 # pyup: != 2.6.0
django-nose==1.4.4
django-test-plus==1.0.18
factory-boy==2.9.2

# pytest
pytest-django==3.1.2
pytest-sugar==0.9.0
pytest-mock==1.6.2
